0.4.4 modify example and add example for zoomIn/zoomOut
